Added libtiff dep check for gppkg
Issue:
if the ibtiff is not available still the postgis gppkg is getting
installed.  this causing the issue while creating a postgis extension
for database.

Fix: added depedency check in .spec file as part of preinstall script.

In gppkg there is plan to add dependency check as part of host
requirement, but that yet to implement. so for now adding check in spec
file to error out if libtiff is not available.