Fix memory corruption issue
Fix issue: https://github.com/apache/cloudberry/issues/1395

In CBDB, pointer is not guaranteed to always be palloc aligned. Due to
our use of MemTuples, the pointer may instead point into the palloc'd
region to an attr offset. Therefore we cannot assume the MemoryContext
from which the pointer was palloc'd exists in the bytes immediately in
front of the pointer.

Instead use MemoryContextContainsGenericAllocation() which correctly
handles the above scenario.